Memperbaiki penyimpanan warna latar Story yang ditolak Postgres
Warna Story disimpan sebagai ARGB. Setiap warna di palet memakai alfa
penuh, sehingga angkanya selalu di atas dua miliar: di luar jangkauan
bilangan bulat empat byte.

Kolomnya sendiri sudah bigint. Yang keliru ada di perintah insert:
nullif($36, 0) membuat Postgres membaca angka 0 sebagai int4 lalu
menyimpulkan tipe yang sama untuk parameternya, dan menolak nilai yang
sah dengan "greater than maximum value for int4". Akibatnya bukan satu
dua warna yang gagal, melainkan seluruh pilihan warna di penyusun Story.

Parameternya kini di-cast ke bigint sebelum dibandingkan.

+++ backend/internal/repository/broadcast.go
@@ -152,7 +152,11 @@ func (r *Repo) SaveBroadcast(
 		        $25, $26, $27, $28, $29,
 		        -- Stored as a bare wall clock; the scheduler reads it in Jakarta.
 		        nullif($30, '')::time, $31, $32, $33, $34, $35,
-		        nullif($36, 0), nullif($37, ''))
+		        -- Cast before comparing. Without it Postgres reads the literal 0 as a
+		        -- four-byte integer and infers the same for the parameter, which
+		        -- refuses every colour whose alpha channel is set: those sit above
+		        -- two billion, which is most of them.
+		        nullif($36::bigint, 0), nullif($37, ''))
 		returning id`,
 		workspaceID, in.ApplicationID, primary, in.CampaignType, in.Name, in.Template,
 		status, in.ScheduledAt, len(in.Targets), createdBy, role, picUser,
